requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Pillow==10.1.0
ultralytics==8.0.227
torch==2.1.1
torchvision==0.16.1